pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "bwf_components"
version = "0.1"
dependencies = [
    "django>=5.1",
    "decouple>=0.0.7",
    "dj-database-url",
    "djangorestframework"
]
description = "A Django app to create flexible, reusable Business Workflows."
readme = "README.rst"
requires-python = ">= 3.10"
authors = [
    {name = "Juan Achury", email = "juan.achurybueno@dbca.wa.gov.au"},
]
classifiers = [
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 5.1",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: BSD License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]
[project.scripts]
"manage.py" = "bwf_components:django_manage"

[project.urls]
Homepage = "https://www.dbca.wa.gov.au/"